+++ node/src/actors/session/handlers.rs
@@ -97,7 +97,7 @@ impl Handler<EpochNotification<EveryEpochPayload>> for Session {
         let current_timestamp = get_timestamp();
         log::trace!(
             "Timestamp diff: {}, Epoch timestamp: {}. Current timestamp: {}",
-            current_timestamp as i64 - msg.timestamp as i64,
+            current_timestamp - msg.timestamp,
             msg.timestamp,
             current_timestamp
         );
