+++ node/src/actors/sessions_manager/handlers.rs
@@ -344,7 +344,7 @@ impl Handler<EpochNotification<()>> for SessionsManager {
         let current_timestamp = get_timestamp();
         log::debug!(
             "Timestamp diff: {}, Epoch timestamp: {}. Current timestamp: {}",
-            current_timestamp as i64 - msg.timestamp as i64,
+            current_timestamp - msg.timestamp,
             msg.timestamp,
             current_timestamp
         );
